Create map.js file and remove map from about.js
Previously, the map on the about page was created using func
createMap(), which generated an iframe that was embedded in a returned
container. Instead, generate the map using the Maps JS API, and create
a function for updating markers after the map has been dragged a
certain amount. Separate these functions into their own module for
better readability.